+++ package.json
@@ -27,8 +27,8 @@
     "u": "node --env-file=.env --no-warnings src/autoshow.js --whisper large --urls",
     "p": "node --env-file=.env --no-warnings src/autoshow.js --whisper large --playlist",
     "f": "node --env-file=.env --no-warnings src/autoshow.js --whisper large --file",
-    "serve": "node --env-file=.env --no-warnings --watch server/index.js",
-    "fetch": "node --env-file=.env --no-warnings server/fetch.js",
+    "serve": "node --env-file=.env --no-warnings --watch packages/server/index.js",
+    "fetch": "node --env-file=.env --no-warnings packages/server/fetch.js",
     "test-local": "node --test test/local.test.js",
     "test-all": "node --test test/all.test.js"
   },
