+++ packages/astro/package.json
