+++ packages/astro/src/components/BaseHead.astro
