+++ packages/astro/src/components/blog/Hero.astro
