+++ packages/astro/src/components/blog/TOC.astro
