+++ packages/astro/src/content/post/2024-07-18-the-navbar-live-prompt.md
