+++ packages/astro/src/pages/index.astro
