+++ packages/astro/src/pages/posts/[...slug].astro
