+++ packages/astro/src/pages/rss.xml.ts
